[ENH] Added SDK download via GitHub Packages Maven
The SA now support download of the Millicast SDK via Maven from GitHub Packages. The earlier method of manually adding the SDK AAR is still supported as before.

To use the Maven download option, ensure:
1) In app/build.gradle dependencies block, "implementation 'com.millicast:millicast-sdk-android:X.X.X'" is available and uncommented, where "X.X.X" is the version of SDK desired, for e.g. "1.1.3".
To use a manually added AAR, comment the above and ensure "implementation project(":MillicastSDK")" is available and uncommented.
2) Credentials for a valid GitHub account is provided, by replacing the values of githubUsername and githubPat in sa.properties with the proper values.